requirements_gui.txt
# Streamlit GUI for GLM-4.6V-Flash Inference
# Install with: pip install -r requirements_gui.txt

# Streamlit framework
streamlit>=1.31.0

# HTTP requests
requests>=2.31.0

# Image processing
Pillow>=10.0.0

# Video processing
opencv-python>=4.8.0
